Fix hidden inputs (#716)
