+++ lib/surface/components/form/hidden_inputs.ex
@@ -17,6 +17,20 @@ defmodule Surface.Components.Form.HiddenInputs do
   prop for, :form, from_context: {Surface.Components.Form, :form}
 
   def render(assigns) do
-    ~F[{hidden_inputs_for(@for)}]
+    ~F"""
+    {#for {name, value_or_values} <- @for.hidden,
+        name = name_for_value_or_values(@for, name, value_or_values),
+        value <- List.wrap(value_or_values)}
+      <input type="hidden" name={name} value={value}>
+    {/for}
+    """
+  end
+
+  defp name_for_value_or_values(form, field, values) when is_list(values) do
+    input_name(form, field) <> "[]"
+  end
+
+  defp name_for_value_or_values(form, field, _value) do
+    input_name(form, field)
   end
 end
